replace implementation of the OCLint report parser
the OCLint report parser implementation was relying on deprecated
functionality, e.g. `StaxParser`, issue builder, etc. and in this
instance it seemed easier to just rewrite rather than to update.

the previous implementation also did not seem to parse a valid or
up-of-date OCLint report, which caused the sensor to not report
anything to sonarqube.